requirements.txt
pydrive2
tqdm
pinecone
python-dotenv
google-genai
groq
PyMuPDF
langchain-text-splitters
sentence-transformers
rank-bm25
